+++ src/action/authorize/paymentMethod/any.ts
@@ -2,10 +2,9 @@ import * as ActionFactory from '../../../action';
 import * as CheckMovieTicketActionFactory from '../../../action/check/paymentMethod/movieTicket';
 import { AvailablePaymentMethodType, IMovieTicket } from '../../../action/trade/pay';
 import { ActionType } from '../../../actionType';
-import * as PayTransactionFactory from '../../../assetTransaction/pay';
+import { ICreditCard, IFromLocation, ITokenizedPaymentCard } from '../../../assetTransaction/pay';
 import { AssetTransactionType } from '../../../assetTransactionType';
 import { IOrderPaymentMethodIssuedThrough, ITotalPaymentDue } from '../../../order';
-import * as CreditCardFactory from '../../../paymentMethod/paymentCard/creditCard';
 import { PaymentStatusType } from '../../../paymentStatusType';
 import { IPropertyValue } from '../../../propertyValue';
 import { TransactionType } from '../../../transactionType';
@@ -27,19 +26,13 @@ export interface IObjectPendingTransaction {
     transactionNumber?: string;
 }
 
-export import ITokenizedPaymentCard = PayTransactionFactory.ITokenizedPaymentCard;
-export import IFromLocation = PayTransactionFactory.IFromLocation;
+export {
+    ICreditCard,
+    IFromLocation,
+    ITokenizedPaymentCard
+};
 
 export import IPurchaseNumberAuthResult = CheckMovieTicketActionFactory.IPurchaseNumberAuthResult;
-
-export import IUnauthorizedCardOfMember = CreditCardFactory.IUnauthorizedCardOfMember;
-export import IUncheckedCardRaw = CreditCardFactory.IUncheckedCardRaw;
-export import IUncheckedCardTokenized = CreditCardFactory.IUncheckedCardTokenized;
-
-/**
- * クレジットカード決済承認アクションに必要なクレジットカード情報
- */
-export type ICreditCard = IUncheckedCardRaw | IUncheckedCardTokenized | IUnauthorizedCardOfMember;
 
 /**
  * 承認対象