+++ src/action/authorize/offer/eventService/coa.ts
@@ -1,4 +1,4 @@
-import * as COA from '@motionpicture/coa-service';
+import type * as COA from '@motionpicture/coa-service';
 
 export interface ICOATicketInfo {
     /**