+++ src/action/transfer/send/order.ts
@@ -5,7 +5,7 @@ import {
     IParticipantAsWebApplication
 } from '../../../action';
 import * as OrderFactory from '../../../order';
-import * as OwnershipInfoFactory from '../../../ownershipInfo';
+import { IGood, IOwnershipInfo } from '../../../ownershipInfo';
 import { IAttributes as IMoneyTransferActionAttributes } from '../../interact/confirm/moneyTransfer';
 import { IAttributes as IRegisterServiceAttributes } from '../../interact/confirm/registerService';
 import * as SendActionFactory from '../send';
@@ -13,13 +13,20 @@ import { IAttributes as ISendEmailMessageActionAttributes } from './message/emai
 
 export type IAgent = IParticipantAsPerson | IParticipantAsProject | IParticipantAsWebApplication;
 export type IRecipient = IParticipantAsWebApplication | IParticipantAsPerson | IParticipantAsCustomer;
-export type IObject = OrderFactory.ISimpleOrder;
+export type IObject = OrderFactory.ISimpleOrder & {
+    /**
+     * 配送対象オファー(2024-01-11~)
+     */
+    acceptedOffers?: {
+        limit: number;
+        page: number;
+    };
+};
 /**
  * 注文配送結果としての所有権
  */
-export type IResult = OwnershipInfoFactory.IOwnershipInfo<OwnershipInfoFactory.IGood>[];
+export type IResult = IOwnershipInfo<IGood>[];
 export interface IPotentialActions {
-    // confirmReservation?: IConfirmReservationActionAttributes<WebAPIFactory.Identifier>[]; // 廃止(2023-08-21~)
     /**
      * 通貨転送アクション
      */