+++ src/creativeWork.ts
@@ -1,4 +1,4 @@
-import * as COA from '@motionpicture/coa-service';
+import type * as COA from '@motionpicture/coa-service';
 
 import { CreativeWorkType } from './creativeWorkType';
 import { IPropertyValue } from './propertyValue';