+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - 注文取引中の外部ロケーションでの決済情報IFを拡張
 - 決済サービスプロバイダーに3Dセキュア認証後加盟店戻りURLを追加
+- 決済サービスプロバイダーに3Dセキュアコールバック方法を追加
 
 ### Changed